requirements.txt
Django==4.2.5
djangorestframework==3.14.0
celery[redis]
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0